+++ lib/crates/aleph-rhi-vulkan/src/descriptor_arena.rs
@@ -94,7 +94,7 @@ impl IDescriptorArena for DescriptorArena {
         };
 
         debug_assert_eq!(sets.len(), sets.capacity());
-        debug_assert_eq!(sets.len(), num_sets as usize);
+        debug_assert_eq!(sets.len(), num_sets);
         unsafe { Ok(core::mem::transmute(sets.into_boxed_slice())) }
     }
